% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PFSFunction.R
\name{runCox}
\alias{runCox}
\title{Identify prognosis-associated Features with Cox Regression}
\usage{
runCox(
  bulk_count,
  bulk_meta,
  design_formula = "Surv(time, status) ~",
  time_event_name = c("time", "status"),
  cv_quantile = 0.05,
  p.adj = 0.05,
  n_cores = parallel::detectCores()
)
}
\arguments{
\item{bulk_count}{A count matrix with features as rows and samples as columns}

\item{bulk_meta}{A data frame containing sample metadata with matching row names to count columns}

\item{design_formula}{design_formula A formula specifying the experimental design (default: "Surv(time, status) ~")}

\item{time_event_name}{the column name for survival time and event status}

\item{cv_quantile}{Filter out a certain proportion of low-variance features by using the coefficient of variation}

\item{p.adj}{Adjusted p-value cutoff for significance (default: 0.05)}

\item{n_cores}{Number of cores for parallel processing. default: parallel::detectCores() (all available cores)}
}
\value{
A list containing:
\itemize{
  \item Cox results for all features
  \item Cox results for significant genes only
}
}
\description{
For survival phenotypes, performs univariate Cox proportional hazards regression on normalized bulk data.
Features with lowest 5% coefficient of variation are excluded to reduce noise. Log2 hazard ratios (HR)
are used as phenotype-specific weights
}
