% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CPscore.R
\name{runMCA}
\alias{runMCA}
\title{Run MCA on Phenotype-specific Feature to Construct Latent Phenotype Space}
\usage{
runMCA(Mtx = Mtx, markers = markers, reduction_name = "mca")
}
\arguments{
\item{Mtx}{A numeric data matrix (features x cells/spots).}

\item{markers}{Character vector of phenotype-specific feature (genes or peaks) to be used.}

\item{reduction_name}{Character, the name assigned to the MCA reduction (default: "mca").}
}
\value{
A list containing:
\itemize{
  \item embeddings: Cell/spot coordinates in MCA space.
  \item features: Features used in the MCA.
}
}
\description{
This function performs Multiple Correspondence Analysis (MCA) on a given
set of phenotype-specific features to generate low-dimensional embeddings of cells/spots.
}
