% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DiPerFunction.R
\name{runPermutationTest}
\alias{runPermutationTest}
\title{Permutation Test for Identifying Phenotype-Associated Cells or spots (CPlink)}
\usage{
runPermutationTest(
  nnSparseMtx,
  posSeedIdx,
  posSeedRWRScores,
  negSeedIdx,
  negSeedRWRScores,
  nPermutations = 100,
  SigThreshold = 0.05,
  nCores = 1,
  restart_prob = 0.05,
  conv_threshold = 1e-05
)
}
\arguments{
\item{nnSparseMtx}{Sparse transition matrix
(cells x cells or spots x spots) used for RWR.}

\item{posSeedIdx}{Integer indices of positive seed nodes.}

\item{posSeedRWRScores}{Numeric vector of diffusion scores from positive seeds.}

\item{negSeedIdx}{Integer indices of negative seed nodes.}

\item{negSeedRWRScores}{Numeric vector of diffusion scores from negative seeds.}

\item{nPermutations}{Number of random permutations to generate the null distribution (default: 100).}

\item{SigThreshold}{Significance threshold for adjusted p-values (default: 0.05).}

\item{nCores}{Number of parallel cores for permutation (default: 1).}

\item{restart_prob}{Restart probability for RWR (default: 0.05).}

\item{conv_threshold}{Convergence threshold for RWR iterations (default: 1e-5).}
}
\value{
A data frame with per-cell or per-spot results.
}
\description{
This function implements the permutation-based statistical framework of CPlink
to assess the significance of Random Walk with Restart (RWR) diffusion scores.
It evaluates whether cells/spots are significantly associated with positive
or negative seed sets compared to a null distribution constructed by random seeds.
}
